+++ internal/wink/end.go
@@ -1,0 +1,3 @@
+package wink
+
+
